Add the logic to save sessin time to the db
In order to calculate the total time spent in a topic, we need to keep track of the start and end time for an exploration and compute the difference between the two, which will give us the time spent in the current session. We will then fetch the previously saved aggregate and its last updated timestamp from TopicLearningTimeDatabase. If the  last updated timestamp is determined to be less than 10 days old, we add the current session time to the aggregate, otherwise we replace it with the current session time, and for both outcomes we update the timestamp.